+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Removed
 
 - `Formal\Migrations\Runner`
+- `Formal\Migrations\Migration`
 
 ## 1.1.0 - 2025-02-28